requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-heroku==0.3.1
django-nose==1.4.6
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
nose==1.3.7
packtools==2.3.8
picles.plumber==0.11
psycopg2==2.7.6.1
pycodestyle==2.4.0
pylint==2.2.2
pytz==2018.7
requests==2.21.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
